libnfs: pread/pwrite: fixes to buffer management
Summary:
pread: return a Vec<u8> rather than taking one as a parameter. The
previous code had a potential buffer overrun (if count was larger than
the passed in buffer.capacity()), and the API was more complex than
necessary. This way pread itself is responsible for managing the
allocation. It does mean each read requires an allocation, but it isn't
that costly compared to an NFS operation.

pwrite: take an immutable buffer reference; write doesn't modify it.

Also infer the pointer types, since it has changed between libnfs
versions.